requirements.txt
PyQt6>=6.6.0
google-generativeai>=0.3.0
loguru>=0.7.0
pytest>=7.4.0
PyQt6-WebEngine>=6.6.0
python-dotenv>=1.0.0